refactor: 접근 제어자 좁힘
